src/nlpia2/ch10/qa_streamlit_app/requirements.txt
streamlit>=1.19.0
transformers
torch
farm-haystack[faiss]
